Restructure app and remove console.log statements. Run linters.
